chore(internal): move resource package to pkg
