+++ pkg/repository/repository.go
@@ -19,9 +19,9 @@ import (
 	"gorm.io/plugin/dbresolver"
 
 	"github.com/instill-ai/model-backend/config"
-	"github.com/instill-ai/model-backend/internal/resource"
 	"github.com/instill-ai/model-backend/pkg/constant"
 	"github.com/instill-ai/model-backend/pkg/datamodel"
+	"github.com/instill-ai/model-backend/pkg/resource"
 	"github.com/instill-ai/x/errmsg"
 	"github.com/instill-ai/x/paginate"
 	"github.com/instill-ai/x/sterr"
