+++ pkg/service/convert.go
@@ -22,9 +22,9 @@ import (
 	"google.golang.org/protobuf/types/known/structpb"
 	"google.golang.org/protobuf/types/known/timestamppb"
 
-	"github.com/instill-ai/model-backend/internal/resource"
 	"github.com/instill-ai/model-backend/pkg/constant"
 	"github.com/instill-ai/model-backend/pkg/datamodel"
+	"github.com/instill-ai/model-backend/pkg/resource"
 
 	custom_logger "github.com/instill-ai/model-backend/pkg/logger"
 	commonpb "github.com/instill-ai/protogen-go/common/task/v1alpha"
