+++ pkg/service/utils.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	"github.com/gofrs/uuid"
-	"github.com/instill-ai/model-backend/internal/resource"
 	"github.com/instill-ai/model-backend/pkg/constant"
+	"github.com/instill-ai/model-backend/pkg/resource"
 )
 
 func (s *service) checkNamespacePermission(ctx context.Context, ns resource.Namespace) error {
